requirements.txt
pytest-playwright==0.7.2
python-dotenv==1.2.1
requests==2.32.5